refactor: streamline imports and update main path resolution in setup and prebuild modules
