+++ template/src/lib/electron-app/factories/app/setup.ts
@@ -1,7 +1,7 @@
 import { app, BrowserWindow } from 'electron'
 
-import { ignoreConsoleWarnings } from '../../utils/utils'
 import { PLATFORM, ENVIRONMENT } from 'shared/constants'
+import { ignoreConsoleWarnings } from '../../utils'
 import { makeAppId } from 'shared/utils'
 
 ignoreConsoleWarnings(['Manifest version 2 is deprecated'])
