+++ template/src/lib/electron-app/utils/index.ts
